requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
gunicorn
python-dotenv
flask-cors